QQ Check Binding/requirements.txt
PyQt5>=5.15.0
requests>=2.31.0
pymysql>=1.1.0
DBUtils>=3.0.3
pyinstaller>=6.3.0
pycryptodome>=3.19.0
psutil>=5.9.0
pywin32>=306
pyarmor>=8.0.1
tinyaes>=1.0.1
wmi>=1.5.1 